工作面板翻译和改进
