+++ VPet-Simulator.Windows/mod/1102_Live2DAnimation
@@ -1,0 +1,1 @@
+D:/Documents/Visual Studio 2022/Projects/VPet.Live2DAnimation/VPet.Plugin.Live2DAnimation/1102_Live2DAnimation
